4.21

+++ part4/blogilista/controllers/blogs.js
@@ -14,6 +14,17 @@ blogsRouter.get('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+
+  const blog = await Blog.findById(request.params.id)
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  if (!decodedToken.id){
+    return response.status(401).json({ error: 'invalid token' })
+  }
+  if (blog.user.toString() !== decodedToken.id) {
+    return response.status(403).json({error: 'only the original poster can delete this post' })
+  }
+
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
